Mostly untested Object support.
Found out from a similar project that the format I'm parsing here is not
HoN-specific, but is in fact simply PHP's serialization format. All is
not lost, since the C version may make a good XS version of
PHP::Serialize on CPAN, but there will have to be some renaming, and the
parser will probably no longer belong to this project.